# xmr/bp: memory optimizations and improvements