… core/src/apps/monero/xmr/bulletproof.py
… core/tests/test_apps.monero.bulletproof.py