Bulletproof implementation optimizations, required for later PR for a private offloading of BP with 16 inputs.

The PR reduces the memory usage with adding only slight time overhead, reduces heap fragmentation (by avoiding unnecessary allocations, using pre-allocated registers). The memory improvement is achieved by on-the-fly vector evaluations wherever possible. 

BP generation:

|BP inputs | Time old [s] | RAM old [B]   | Time new [s] | RAM new [B] |
|----------|---------|------------|-----------|----------|
|1               | 15.70      | 12 848       | 16.88         | 9 648     |
|2               | 28.00     | 21 408      | 30.55         | 13 536   |
|4               | 52.17      | 38 304     | 57.40          | 22 224   |
|8               | 107.46    | 78 592      | 121.21       | 39 696    |
|16             | FAIL        | > 89 120   | 246.57      | 74 400   |

BP verification:

|BP inputs | Time old [s] | RAM old [B]   | Time new [s] | RAM new [B] |
|----------|---------|------------|-----------|----------|
|1               | 5.12         | 5 856        | 5.12           | 5 664     |
|2               | 9.60        | 6 480        | 9.58          | 6 256     |
|4               | 18.56      | 7 168         | 18.56        | 6 912      |
|8               | 39.11       | 7 904        | 39.00       | 7 616      |
|16             | -              | -                | 80.40       | 8 512     |

Changes:
- different approach to vector manipulation - more on the fly operations. Prepared for fully offloaded operations, BP on Trezor with constant memory.
- memory requirements reduced from (4MN + 2lg(MN) + const) to (2MN +  2lg(MN) + const)
- more raw methods to avoid unnecessary encoding/decoding
- chunking improved, chunk size set as a constant, changed from 64 to 32, missing pieces implemented to cover also BP 16
- proof_v8 support discontinued, old hardfork, not needed anymore
- get_exponent register clash fixed (for large vectors)
- reduced heap fragmentation by removing some temporary allocations
- hashing with len and offset to reduce heap fragmentation by creating a sliced arrays
- use to() wherever possible to avoid allocations and return of mutable private object
- global functions start with _ prefix, reduce import footprint
- use `__slots__` in classes to minimize footprint