@tsusanka 2019-11-20
> global functions start with _ prefix, reduce import footprint

Didn't you mean _local_ functions?