@ph4r05 2019-11-20
> > global functions start with _ prefix, reduce import footprint
> 
> Didn't you mean _local_ functions?

Yeah I mean function on the module scope.