@ph4r05 2020-01-13
BTW - I submitted a paper on a offloaded BP computation. I've designed the MPC BP prover, which has the following complexity:

![bp-mpc-timing](https://user-images.githubusercontent.com/1052761/72267989-4785e780-3621-11ea-92c9-5ad065b908ac.png)

(Sorry for just image, it is in LaTeX and this is the fastest way to convey the stats now)
I can send you pre-print privately if you are interested in this. I have already the implementation for Trezor ready. It could be PRed.