@onvej-sl 2020-02-20
@ph4r05 

I rebased your PR, it can be found [here](https://github.com/trezor/trezor-firmware/tree/onvej-sl/monero-optimizations). I want you to confirm that you agree with the changes I made. The first four commits just fix style that is newly enforced.

The commit "Fix flake undefined variables warning II" suppresses flake messages that warn you against the following situation, that in my opinion doesn't occur in the code:
```
class B:
    def __init__(self, c):
        self.c = c

a = 0
b = B(lambda: a)
del a
b.c() #NameError: name 'a' is not defined
```

In the commit "Fix flake undefined variables warning II" I made `_CHSIZE` and `_CHBITS` (class attribute) variables global since:
  * flake is complaining that the variables are undefined and
  * the constant variables in micropython are global anyway.